Add form validation toast
