+++ src/components/Process/Chained.tsx
@@ -6,15 +6,15 @@ import {
   ElectionResults,
   QuestionsFormProvider,
   SpreadsheetAccess,
-  SubmitFormValidation,
 } from '@vocdoni/chakra-components'
 import { ElectionProvider, useElection } from '@vocdoni/react-providers'
 import { InvalidElection, IVotePackage, PublishedElection, VocdoniSDKClient } from '@vocdoni/sdk'
 import { PropsWithChildren, useEffect, useState } from 'react'
-import { Trans, useTranslation } from 'react-i18next'
+import { Trans } from 'react-i18next'
 import { VoteButton } from '~components/Process/Aside'
 import BlindCSPConnect from '~components/Process/BlindCSPConnect'
 import { ChainedProvider, useChainedProcesses } from './ChainedContext'
+import { useFormValidation } from '~components/Process/ParitaryErc'
 
 type ChainedProcessesInnerProps = {
   connected: boolean
@@ -35,9 +35,9 @@ const VoteButtonContainer = ({ children }: PropsWithChildren) => {
 }
 
 const ChainedProcessesInner = ({ connected }: ChainedProcessesInnerProps) => {
-  const { t } = useTranslation()
   const { election, voted, setClient, clearClient } = useElection()
   const { processes, client, current, setProcess, setCurrent } = useChainedProcesses()
+  const { formValidation } = useFormValidation()
 
   // clear session of local context when login out
   useEffect(() => {
@@ -90,12 +90,6 @@ const ChainedProcessesInner = ({ connected }: ChainedProcessesInnerProps) => {
   }
 
   if (isRenderWith) {
-    const formValidation: SubmitFormValidation = (values) => {
-      if (!sameLengthValidator(values)) {
-        return t('errors.all_ballots_must_have_same_length')
-      }
-      return true
-    }
     return (
       <QuestionsFormProvider validate={formValidation} renderWith={renderWith}>
         <ElectionQuestionsForm />
@@ -361,16 +355,4 @@ export type FlowNode =
 
 export type RenderWith = {
   id: string
-}
-
-/**
- * Check all values responses have the same length
- * Won't work for multiquestions elections.
- */
-export const sameLengthValidator: SubmitFormValidation = (answers) => {
-  const [first, ...rest] = Object.values(answers)
-  if (!first) {
-    throw new Error('No fields found')
-  }
-  return rest.every((ballot) => ballot[0].length === first[0].length)
 }
